pydantic 2.8.2, pydantic-settings 2.3.4
