+++ CodeGenerationTools/GridworksCore/Enum/Docs/DeriveAslDocs.xslt
@@ -67,7 +67,8 @@
     <xsl:text>"</xsl:text>
     <xsl:value-of select="LocalValue"/>
     <xsl:text>"</xsl:text>
-        <xsl:if test="position() != count($airtable//EnumSymbols/EnumSymbol[(Enum = $enum-id) and (Version &lt;= $version)])">
+
+
         <xsl:text>, </xsl:text>
     </xsl:if>
