+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "gridworks-base"
-version = "0.0.3"
+version = "0.1.0"
 description = "Gridworks Base"
 authors = ["GridWorks <gridworks@gridworks-consulting.com>"]
 license = "MIT"
@@ -23,13 +23,14 @@ Changelog = "https://github.com/thegridelectric/gridworks-base/releases"
 [tool.poetry.dependencies]
 python = ">=3.11, <4.0"
 python-dotenv = ">=1.0.0"
-pydantic = ">=1.10.2"
 #pydantic = ">=2.8.2"
 
 pika = "^1.3.2"
 py-algorand-sdk = ">=2.6.1"
+pydantic = "^2.8.2"
+pydantic-settings = "^2.3.4"
+gridworks = "^1.1.6"
 #gridworks = {path="../gridworks", develop=true}
-gridworks = "^1.1.5"
 
 [tool.poetry.dev-dependencies]
 Pygments = ">=2.10.0"
