+++ src/gwbase/config/__init__.py
@@ -1,0 +1,10 @@
+""" Base GNodeSettings """
+
+from gwbase.config.g_node_settings import GNodeSettings
+from gwbase.config.g_node_settings import SupervisorSettings
+
+
+__all__ = [
+    "GNodeSettings",
+    "SupervisorSettings",
+]
