+++ src/gwbase/types/base_g_node_gt.py
@@ -8,11 +8,12 @@ from typing import Literal
 from typing import Optional
 
 from gw.errors import GwTypeError
-from gw.utils import snake_to_camel
+from gw.utils import is_pascal_case
 from pydantic import BaseModel
-from pydantic import Extra
 from pydantic import Field
-from pydantic import validator
+from pydantic import field_validator
+from pydantic.alias_generators import to_pascal
+from pydantic.alias_generators import to_snake
 
 from gwbase.data_classes.base_g_node import BaseGNode
 from gwbase.enums import CoreGNodeRole
@@ -87,11 +88,11 @@ class BaseGNodeGt(BaseModel):
     version: Literal["002"] = "002"
 
     class Config:
-        extra = Extra.allow
-        allow_population_by_field_name = True
-        alias_generator = snake_to_camel
+        extra = "allow"
+        populate_by_name = True
+        alias_generator = to_pascal
 
-    @validator("g_node_id")
+    @field_validator("g_node_id")
     def _check_g_node_id(cls, v: str) -> str:
         try:
             check_is_uuid_canonical_textual(v)
@@ -101,7 +102,7 @@ class BaseGNodeGt(BaseModel):
             )
         return v
 
-    @validator("alias")
+    @field_validator("alias")
     def _check_alias(cls, v: str) -> str:
         try:
             check_is_left_right_dot(v)
@@ -109,7 +110,7 @@ class BaseGNodeGt(BaseModel):
             raise ValueError(f"Alias failed LeftRightDot format validation: {e}")
         return v
 
-    @validator("g_node_registry_addr")
+    @field_validator("g_node_registry_addr")
     def _check_g_node_registry_addr(cls, v: str) -> str:
         try:
             check_is_algo_address_string_format(v)
@@ -119,7 +120,7 @@ class BaseGNodeGt(BaseModel):
             )
         return v
 
-    @validator("prev_alias")
+    @field_validator("prev_alias")
     def _check_prev_alias(cls, v: Optional[str]) -> Optional[str]:
         if v is None:
             return v
@@ -129,7 +130,7 @@ class BaseGNodeGt(BaseModel):
             raise ValueError(f"PrevAlias failed LeftRightDot format validation: {e}")
         return v
 
-    @validator("gps_point_id")
+    @field_validator("gps_point_id")
     def _check_gps_point_id(cls, v: Optional[str]) -> Optional[str]:
         if v is None:
             return v
@@ -141,7 +142,7 @@ class BaseGNodeGt(BaseModel):
             )
         return v
 
-    @validator("ownership_deed_id")
+    @field_validator("ownership_deed_id")
     def _check_ownership_deed_id(cls, v: Optional[int]) -> Optional[int]:
         if v is None:
             return v
@@ -153,7 +154,7 @@ class BaseGNodeGt(BaseModel):
             )
         return v
 
-    @validator("ownership_deed_validator_addr")
+    @field_validator("ownership_deed_validator_addr")
     def _check_ownership_deed_validator_addr(cls, v: Optional[str]) -> Optional[str]:
         if v is None:
             return v
@@ -165,7 +166,7 @@ class BaseGNodeGt(BaseModel):
             )
         return v
 
-    @validator("owner_addr")
+    @field_validator("owner_addr")
     def _check_owner_addr(cls, v: Optional[str]) -> Optional[str]:
         if v is None:
             return v
@@ -177,7 +178,7 @@ class BaseGNodeGt(BaseModel):
             )
         return v
 
-    @validator("daemon_addr")
+    @field_validator("daemon_addr")
     def _check_daemon_addr(cls, v: Optional[str]) -> Optional[str]:
         if v is None:
             return v
@@ -189,7 +190,7 @@ class BaseGNodeGt(BaseModel):
             )
         return v
 
-    @validator("trading_rights_id")
+    @field_validator("trading_rights_id")
     def _check_trading_rights_id(cls, v: Optional[int]) -> Optional[int]:
         if v is None:
             return v
@@ -201,7 +202,7 @@ class BaseGNodeGt(BaseModel):
             )
         return v
 
-    @validator("scada_algo_addr")
+    @field_validator("scada_algo_addr")
     def _check_scada_algo_addr(cls, v: Optional[str]) -> Optional[str]:
         if v is None:
             return v
@@ -213,7 +214,7 @@ class BaseGNodeGt(BaseModel):
             )
         return v
 
-    @validator("scada_cert_id")
+    @field_validator("scada_cert_id")
     def _check_scada_cert_id(cls, v: Optional[int]) -> Optional[int]:
         if v is None:
             return v
@@ -242,10 +243,8 @@ class BaseGNodeGt(BaseModel):
         It also applies these changes recursively to sub-types.
         """
         d = {
-            key: value
-            for key, value in self.dict(
-                by_alias=True, include=self.__fields_set__ | {"type_name", "version"}
-            ).items()
+            to_pascal(key): value
+            for key, value in self.model_dump().items()
             if value is not None
         }
         del d["Status"]
@@ -330,6 +329,9 @@ class BaseGNodeGt_Maker:
         Returns:
             BaseGNodeGt
         """
+        for key in d.keys():
+            if not is_pascal_case(key):
+                raise GwTypeError(f"Key '{key}' is not PascalCase")
         d2 = dict(d)
         if "GNodeId" not in d2.keys():
             raise GwTypeError(f"dict missing GNodeId: <{d2}>")
@@ -356,7 +358,8 @@ class BaseGNodeGt_Maker:
                 f"Attempting to interpret base.g.node.gt version {d2['Version']} as version 002"
             )
             d2["Version"] = "002"
-        return BaseGNodeGt(**d2)
+        d3 = {to_snake(key): value for key, value in d2.items()}
+        return BaseGNodeGt(**d3)
 
     @classmethod
     def tuple_to_dc(cls, t: BaseGNodeGt) -> BaseGNode:
