+++ src/gwbase/types/heartbeat_a.py
@@ -7,10 +7,12 @@ from typing import Dict
 from typing import Literal
 
 from gw.errors import GwTypeError
-from gw.utils import snake_to_camel
+from gw.utils import is_pascal_case
 from pydantic import BaseModel
 from pydantic import Field
-from pydantic import validator
+from pydantic import field_validator
+from pydantic.alias_generators import to_pascal
+from pydantic.alias_generators import to_snake
 
 
 LOG_FORMAT = (
@@ -42,10 +44,10 @@ class HeartbeatA(BaseModel):
     version: Literal["100"] = "100"
 
     class Config:
-        allow_population_by_field_name = True
-        alias_generator = snake_to_camel
+        populate_by_name = True
+        alias_generator = to_pascal
 
-    @validator("my_hex")
+    @field_validator("my_hex")
     def _check_my_hex(cls, v: str) -> str:
         try:
             check_is_hex_char(v)
@@ -53,7 +55,7 @@ class HeartbeatA(BaseModel):
             raise ValueError(f"MyHex failed HexChar format validation: {e}")
         return v
 
-    @validator("your_last_hex")
+    @field_validator("your_last_hex")
     def _check_your_last_hex(cls, v: str) -> str:
         try:
             check_is_hex_char(v)
@@ -78,10 +80,8 @@ class HeartbeatA(BaseModel):
         It also applies these changes recursively to sub-types.
         """
         d = {
-            key: value
-            for key, value in self.dict(
-                by_alias=True, include=self.__fields_set__ | {"type_name", "version"}
-            ).items()
+            to_pascal(key): value
+            for key, value in self.model_dump().items()
             if value is not None
         }
         return d
@@ -162,6 +162,9 @@ class HeartbeatA_Maker:
         Returns:
             HeartbeatA
         """
+        for key in d.keys():
+            if not is_pascal_case(key):
+                raise GwTypeError(f"Key '{key}' is not PascalCase")
         d2 = dict(d)
         if "MyHex" not in d2.keys():
             raise GwTypeError(f"dict missing MyHex: <{d2}>")
@@ -176,7 +179,8 @@ class HeartbeatA_Maker:
                 f"Attempting to interpret heartbeat.a version {d2['Version']} as version 100"
             )
             d2["Version"] = "100"
-        return HeartbeatA(**d2)
+        d3 = {to_snake(key): value for key, value in d2.items()}
+        return HeartbeatA(**d3)
 
 
 def check_is_hex_char(v: str) -> None:
