+++ src/gwbase/types/ready.py
@@ -7,10 +7,12 @@ from typing import Dict
 from typing import Literal
 
 from gw.errors import GwTypeError
-from gw.utils import snake_to_camel
+from gw.utils import is_pascal_case
 from pydantic import BaseModel
 from pydantic import Field
-from pydantic import validator
+from pydantic import field_validator
+from pydantic.alias_generators import to_pascal
+from pydantic.alias_generators import to_snake
 
 
 LOG_FORMAT = (
@@ -48,10 +50,10 @@ class Ready(BaseModel):
     version: Literal["001"] = "001"
 
     class Config:
-        allow_population_by_field_name = True
-        alias_generator = snake_to_camel
+        populate_by_name = True
+        alias_generator = to_pascal
 
-    @validator("from_g_node_alias")
+    @field_validator("from_g_node_alias")
     def _check_from_g_node_alias(cls, v: str) -> str:
         try:
             check_is_left_right_dot(v)
@@ -61,7 +63,7 @@ class Ready(BaseModel):
             )
         return v
 
-    @validator("from_g_node_instance_id")
+    @field_validator("from_g_node_instance_id")
     def _check_from_g_node_instance_id(cls, v: str) -> str:
         try:
             check_is_uuid_canonical_textual(v)
@@ -88,10 +90,8 @@ class Ready(BaseModel):
         It also applies these changes recursively to sub-types.
         """
         d = {
-            key: value
-            for key, value in self.dict(
-                by_alias=True, include=self.__fields_set__ | {"type_name", "version"}
-            ).items()
+            to_pascal(key): value
+            for key, value in self.model_dump().items()
             if value is not None
         }
         return d
@@ -172,6 +172,9 @@ class Ready_Maker:
         Returns:
             Ready
         """
+        for key in d.keys():
+            if not is_pascal_case(key):
+                raise GwTypeError(f"Key '{key}' is not PascalCase")
         d2 = dict(d)
         if "FromGNodeAlias" not in d2.keys():
             raise GwTypeError(f"dict missing FromGNodeAlias: <{d2}>")
@@ -188,7 +191,8 @@ class Ready_Maker:
                 f"Attempting to interpret ready version {d2['Version']} as version 001"
             )
             d2["Version"] = "001"
-        return Ready(**d2)
+        d3 = {to_snake(key): value for key, value in d2.items()}
+        return Ready(**d3)
 
 
 def check_is_left_right_dot(v: str) -> None:
