+++ src/gwbase/types/supervisor_container_gt.py
@@ -7,11 +7,12 @@ from typing import Dict
 from typing import Literal
 
 from gw.errors import GwTypeError
-from gw.utils import snake_to_camel
+from gw.utils import is_pascal_case
 from pydantic import BaseModel
-from pydantic import Extra
 from pydantic import Field
-from pydantic import validator
+from pydantic import field_validator
+from pydantic.alias_generators import to_pascal
+from pydantic.alias_generators import to_snake
 
 from gwbase.enums import SupervisorContainerStatus
 
@@ -55,11 +56,11 @@ class SupervisorContainerGt(BaseModel):
     version: Literal["000"] = "000"
 
     class Config:
-        extra = Extra.allow
-        allow_population_by_field_name = True
-        alias_generator = snake_to_camel
+        extra = "allow"
+        populate_by_name = True
+        alias_generator = to_pascal
 
-    @validator("supervisor_container_id")
+    @field_validator("supervisor_container_id")
     def _check_supervisor_container_id(cls, v: str) -> str:
         try:
             check_is_uuid_canonical_textual(v)
@@ -69,7 +70,7 @@ class SupervisorContainerGt(BaseModel):
             )
         return v
 
-    @validator("world_instance_name")
+    @field_validator("world_instance_name")
     def _check_world_instance_name(cls, v: str) -> str:
         try:
             check_is_world_instance_name_format(v)
@@ -79,7 +80,7 @@ class SupervisorContainerGt(BaseModel):
             )
         return v
 
-    @validator("supervisor_g_node_instance_id")
+    @field_validator("supervisor_g_node_instance_id")
     def _check_supervisor_g_node_instance_id(cls, v: str) -> str:
         try:
             check_is_uuid_canonical_textual(v)
@@ -89,7 +90,7 @@ class SupervisorContainerGt(BaseModel):
             )
         return v
 
-    @validator("supervisor_g_node_alias")
+    @field_validator("supervisor_g_node_alias")
     def _check_supervisor_g_node_alias(cls, v: str) -> str:
         try:
             check_is_left_right_dot(v)
@@ -116,10 +117,8 @@ class SupervisorContainerGt(BaseModel):
         It also applies these changes recursively to sub-types.
         """
         d = {
-            key: value
-            for key, value in self.dict(
-                by_alias=True, include=self.__fields_set__ | {"type_name", "version"}
-            ).items()
+            to_pascal(key): value
+            for key, value in self.model_dump().items()
             if value is not None
         }
         del d["Status"]
@@ -202,6 +201,9 @@ class SupervisorContainerGt_Maker:
         Returns:
             SupervisorContainerGt
         """
+        for key in d.keys():
+            if not is_pascal_case(key):
+                raise GwTypeError(f"Key '{key}' is not PascalCase")
         d2 = dict(d)
         if "SupervisorContainerId" not in d2.keys():
             raise GwTypeError(f"dict missing SupervisorContainerId: <{d2}>")
@@ -225,7 +227,8 @@ class SupervisorContainerGt_Maker:
                 f"Attempting to interpret supervisor.container.gt version {d2['Version']} as version 000"
             )
             d2["Version"] = "000"
-        return SupervisorContainerGt(**d2)
+        d3 = {to_snake(key): value for key, value in d2.items()}
+        return SupervisorContainerGt(**d3)
 
 
 def check_is_left_right_dot(v: str) -> None:
